+++ src/test/java/dev/orne/beans/ValidIdentityIT.java
@@ -107,7 +107,7 @@ class ValidIdentityIT {
      */
     @Test
     void testValidTokenValidationNullBody() {
-        final String token = "ID" + IdentityTokenFormatter.NULL_TOKEN;
+        final String token = "ID" + IdentityTokenFormatter.NULL_BODY;
         final TestContainer nullableContainer = new TestContainer();
         nullableContainer.token = token;
         assertTrue(BeanValidationUtils.isValid(nullableContainer));